Moved to models to avoid circular imports
